feat: Add UserSetting Prisma model
This commit introduces the UserSetting Prisma model, representing user-specific settings in the application. The model includes fields such as id, userId (linked to the User model), isDarkMode, numberOfColumns, showNumberOfBookmarkInTab, showNumberOfBookmarkInCategory, showTagsInTooltip, showNoteInTooltip, allowDragDropToMoveLink, createdAt, and updatedAt.

The 'ColumnNumber' enum is utilized to define the possible values for the 'numberOfColumns' field, providing a structured and predefined set of options. The model is designed to capture user preferences and configurations, contributing to a personalized user experience.

This addition sets the groundwork for storing and managing user-specific settings in the application database.